# openapi_rapidapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EU AI Act Compliance Scanner API",
    "description": "Scan GitHub repos or source code for EU AI Act compliance. Detects 16+ AI frameworks, checks compliance, provides actionable recommendations. Free: 5 scans/day. API key: unlimited.",
    "version": "1.0.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/scan": {
      "post": {
        "summary": "Scan",
        "description": "Scan a GitHub repo URL or inline source code for EU AI Act compliance.\n\nProvide either `url` (GitHub HTTPS URL) or `code_source` (inline code).\nReturns risk score, detected frameworks, compliance score, and recommendations.",
        "operationId": "scan_v1_scan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/scan/{scan_id}": {
      "get": {
        "summary": "Get Scan",
        "description": "Retrieve a previous scan result by ID.",
        "operationId": "get_scan_v1_scan__scan_id__get",
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/usage": {
      "get": {
        "summary": "Usage",
        "description": "Check your current rate limit usage.",
        "operationId": "usage_v1_usage_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/scan/async": {
      "post": {
        "summary": "Scan Async",
        "description": "Start an async scan job. Returns immediately with a job_id.\n\nUse GET /v1/job/{job_id} to poll for results.\nUseful for large repos that may exceed the 60s synchronous timeout.",
        "operationId": "scan_async_v1_scan_async_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/job/{job_id}": {
      "get": {
        "summary": "Get Job",
        "description": "Poll an async scan job status.\n\nStatus values: pending \u2192 running \u2192 done | error\nWhen status is 'done', the full scan result is in the 'result' field.",
        "operationId": "get_job_v1_job__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ScanRequest": {
        "properties": {
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url",
            "description": "GitHub repo URL to scan (HTTPS)"
          },
          "code_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code Source",
            "description": "Inline source code to scan"
          },
          "risk_category": {
            "type": "string",
            "title": "Risk Category",
            "description": "Risk category: unacceptable, high, limited, minimal",
            "default": "limited"
          }
        },
        "type": "object",
        "title": "ScanRequest"
      },
      "ScanResponse": {
        "properties": {
          "scan_id": {
            "type": "string",
            "title": "Scan Id"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "risk_score": {
            "type": "number",
            "title": "Risk Score"
          },
          "risk_category": {
            "type": "string",
            "title": "Risk Category"
          },
          "frameworks_detected": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Frameworks Detected"
          },
          "files_scanned": {
            "type": "integer",
            "title": "Files Scanned"
          },
          "compliance_score": {
            "type": "string",
            "title": "Compliance Score"
          },
          "recommendations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Recommendations"
          },
          "source": {
            "type": "string",
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "scan_id",
          "timestamp",
          "risk_score",
          "risk_category",
          "frameworks_detected",
          "files_scanned",
          "compliance_score",
          "recommendations",
          "source"
        ],
        "title": "ScanResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "servers": [
    {
      "url": "https://mcp.arkforge.fr/api",
      "description": "Production"
    },
    {
      "url": "http://localhost:8200",
      "description": "Local development"
    }
  ]
}